feat(Telegram): extend Message and User records to include InlineKeyboardMarkup and change Id type

+++ Services/TelegramBot/Types/User.cs
@@ -1,3 +1,9 @@
 namespace TelegramBot.Types;
 
-internal record class User(int Id, bool IsBot, string FirstName, string? LastName, string? Username);
+internal record class User(
+    long Id,
+    bool IsBot,
+    string FirstName,
+    string? LastName,
+    string? Username
+);
